refactor: group textfield & checkbox in forms
