+++ src/new-components/forms/checkbox/checkbox.stories.mdx
@@ -1,0 +1,20 @@
+import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
+import { Checkbox } from './checkbox'; 
+
+<Meta
+  argTypes={{onChange: { action: 'onChange' }}}
+  title="Forms/Checkbox"
+  component={{ Checkbox }}/>
+
+> ⚠️ API not usable yet, design focused only
+
+# Checkbox
+Checkboxes are used for a list of options where the user may select multiple options, including all or none.
+
+<Canvas withSource='open'>
+  <Story name="Regular">
+    <Checkbox label='Satellytes'/>
+  </Story>
+</Canvas>
+
+<ArgsTable of={Checkbox}/>
