+++ src/new-components/forms/checkbox/checkbox.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import { TextStyles } from '../typography/typography-v2';
+import { TextStyles } from '../../../components/typography/typography-v2';
 
-const Checkbox = (props) => {
+export const Checkbox = (props) => {
   const [checked, setChecked] = useState(true);
 
   const handleChange = () => {
@@ -71,5 +71,3 @@ const Checkbox = (props) => {
     </CheckboxLabel>
   );
 };
-
-export default Checkbox;
