+++ src/new-components/forms/textinput/textinput.stories.mdx
@@ -1,0 +1,18 @@
+import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs'; 
+import { TextInput } from './textinput';
+
+<Meta
+  title="Forms/TextInput"
+  component={{ TextInput }}/>
+
+> ⚠️ API not usable yet, design focused only
+
+# TextInput
+
+<Canvas withSource='open'>
+  <Story name='Regular'>
+    <TextInput label="First Name *"/>
+  </Story>
+</Canvas>
+
+<ArgsTable of={TextInput}/>
